move all types into own module
