+++ src/types/binding.rs
@@ -2,18 +2,20 @@ pub mod py_centerline;
 pub mod py_centerline_point;
 pub mod py_contour;
 pub mod py_contour_point;
+pub mod py_discretized_vessel_tree;
 pub mod py_frame;
 pub mod py_geometry;
 pub mod py_geometry_pair;
 pub mod py_input_data;
-pub mod record;
+pub mod py_record;
 
 pub use py_centerline::PyCenterline;
 pub use py_centerline_point::PyCenterlinePoint;
 pub use py_contour::{PyContour, PyContourType};
 pub use py_contour_point::PyContourPoint;
+pub use py_discretized_vessel_tree::PyDiscretizedVesselTree;
 pub use py_frame::PyFrame;
 pub use py_geometry::PyGeometry;
 pub use py_geometry_pair::PyGeometryPair;
 pub use py_input_data::PyInputData;
-pub use record::PyRecord;
+pub use py_record::PyRecord;
