+++ src/types/native.rs
@@ -2,6 +2,7 @@ pub mod centerline;
 pub mod centerline_point;
 pub mod contour;
 pub mod contour_point;
+pub mod discretized_tree;
 pub mod frame;
 pub mod geometry;
 pub mod geometry_pair;
@@ -11,7 +12,8 @@ pub mod record;
 pub use centerline::Centerline;
 pub use centerline_point::CenterlinePoint;
 pub use contour::{Contour, ContourType};
-pub use contour_point::{downsample_contour_points, ContourPoint};
+pub use contour_point::{downsample_contour_points, ContourPoint, Point3D};
+pub use discretized_tree::{DiscretizedVesselTree, ReferenceTriplet};
 pub use frame::Frame;
 pub use geometry::Geometry;
 pub use geometry_pair::GeometryPair;
